define RestartCategoriesConfig only in fv3kube
