+++ external/fv3kube/fv3kube/config.py
@@ -2,7 +2,6 @@ import os
 import datetime
 from pathlib import Path
 from typing import Any, Sequence, Mapping, Optional
-from typing_extensions import Protocol
 import dataclasses
 import fsspec
 import yaml
@@ -19,12 +18,6 @@ BASE_FV3CONFIG_BY_VERSION = {
     "v0.5": os.path.join(PWD, "base_yamls/v0.5/fv3config.yml"),
     "v0.6": os.path.join(PWD, "base_yamls/v0.6/fv3config.yml"),
 }
-STANDARD_RESTART_CATEGORIES = {
-    "core": "fv_core.res",
-    "surface": "sfc_data",
-    "tracer": "fv_tracer.res",
-    "surface_wind": "fv_srf_wnd.res",
-}
 TILE_COORDS_FILENAMES = range(1, 7)  # tile numbering in model output filenames
 FV_CORE_ASSET = fv3config.get_asset_dict(
     "gs://vcm-fv3config/data/initial_conditions/fv_core_79_levels/v1.0/",
@@ -32,6 +25,23 @@ FV_CORE_ASSET = fv3config.get_asset_dict(
     target_location="INPUT",
 )
 FV3Config = Mapping[str, Any]
+
+
+@dataclasses.dataclass
+class RestartCategoriesConfig:
+    """A specification of the restart file names by category
+
+    Attributes:
+        core: name of fv3 core restart files
+        surface: name of surface restart files
+        tracer: name of tracer restart files
+        surface_wind: name of surface wind restart files
+    """
+
+    core: str = "fv_core.res"
+    surface: str = "sfc_data"
+    tracer: str = "fv_tracer.res"
+    surface_wind: str = "fv_srf_wnd.res"
 
 
 def _merge_once(source, update):
@@ -82,13 +92,6 @@ def get_base_fv3config(version_key: Optional[str] = None) -> FV3Config:
     return base_yaml
 
 
-class RestartCategoriesConfig(Protocol):
-    core: str
-    surface: str
-    tracer: str
-    surface_wind: str
-
-
 def update_tiled_asset_names(
     source_url: str,
     source_filename: str,
@@ -106,23 +109,24 @@ def update_tiled_asset_names(
     Filename strings should include any specified variable name inserts to
     be updated with a format. E.g., "{timestep}.{category}.tile{tile}.nc"
     """
+    standard_restart_categories = RestartCategoriesConfig()
     if restart_categories is None:
-        restart_categories = STANDARD_RESTART_CATEGORIES
-    else:
-        restart_categories = dataclasses.asdict(restart_categories)
+        restart_categories = RestartCategoriesConfig()
 
     assets = [
         fv3config.get_asset_dict(
             source_url,
             source_filename.format(
-                category=restart_categories[category_name], tile=tile, **kwargs
+                category=getattr(restart_categories, category_name), tile=tile, **kwargs
             ),
             target_location=target_url,
             target_name=target_filename.format(
-                category=STANDARD_RESTART_CATEGORIES[category_name], tile=tile, **kwargs
+                category=getattr(standard_restart_categories, category_name),
+                tile=tile,
+                **kwargs,
             ),
         )
-        for category_name in restart_categories.keys()
+        for category_name in vars(restart_categories)
         for tile in TILE_COORDS_FILENAMES
     ]
